minor update to install script

+++ install.sh
@@ -12,17 +12,21 @@ cd $HOME
 
 sudo rm -rf $HOME/homebrew/plugins/LegionGoRemapper
 
+if [ ! -f "/etc/udev/rules.d/90-legion-go-remapper.rules" ]; then
+
 cat << EOF | sudo tee -a "/etc/udev/rules.d/90-legion-go-remapper.rules"
 # allow r/w access by all local/physical sessions (seats)
 # https://github.com/systemd/systemd/issues/4288
 SUBSYSTEMS=="usb", ATTRS{idVendor}=="17ef", TAG+="uaccess"
 
 # allow r/w access by users of the plugdev group
-SUBSYSTEMS=="usb", ATTRS{idVendor}=="17ef", GROUP="plugdev", MODE="0660"
+# SUBSYSTEMS=="usb", ATTRS{idVendor}=="17ef", GROUP="plugdev", MODE="0660"
 
 # allow r/w access by all users
 SUBSYSTEMS=="usb", ATTRS{idVendor}=="17ef", MODE="0666"
 EOF
+
+fi
 
 echo "installing LegionGoRemapper plugin for RGB control"
 # download + install Legion go remapper
